made dumper a relay instead of a pneumatic.  Dumper still has no timers.  For now, I figure you just put 3 buttons on a controller for on, off, and reverse and eyeball it on the pyramid.
Added indexer and OI button where Kidd wants it.  Shooter still needs to be added.  Index time needs to be tuned in RobotMap.  I couldn't remember if extending or retracting the pneumatic pushes a frisbee, so the names may be screwed up, but it should work fine, theres a bunch of places you can change it.